Fix migration
