+++ alembic/versions/20250408_104027_fc2a6e745865_default_migration_message.py
@@ -1,22 +1,21 @@
 """Default migration message
 
-Revision ID: d56333140172
+Revision ID: fc2a6e745865
 Revises: 0f59f5993a23
-Create Date: 2025-04-14 15:19:26.834055
+Create Date: 2025-04-08 10:40:27.097324
 
 """
 
 from collections.abc import Sequence
 
 import sqlalchemy as sa
-from alembic_postgresql_enum import TableReference
 from alembic_utils.pg_trigger import PGTrigger
 from sqlalchemy.dialects import postgresql
 
 from alembic import op
 
 # revision identifiers, used by Alembic.
-revision: str = "d56333140172"
+revision: str = "fc2a6e745865"
 down_revision: str | None = "0f59f5993a23"
 branch_labels: str | Sequence[str] | None = None
 depends_on: str | Sequence[str] | None = None
@@ -24,8 +23,6 @@ depends_on: str | Sequence[str] | None = None
 
 def upgrade() -> None:
     # ### commands auto generated by Alembic - please adjust! ###
-    sa.Enum("male", "female", name="sex").create(op.get_bind())
-    sa.Enum("prenatal", "postnatal", name="ageperiod").create(op.get_bind())
     op.add_column(
         "analysis_software_source_code",
         sa.Column("description_vector", postgresql.TSVECTOR(), nullable=True),
@@ -112,86 +109,14 @@ def upgrade() -> None:
         ["name"],
         unique=False,
     )
-    op.add_column("subject", sa.Column("age_value", sa.Interval(), nullable=True))
-    op.add_column("subject", sa.Column("age_min", sa.Interval(), nullable=True))
-    op.add_column("subject", sa.Column("age_max", sa.Interval(), nullable=True))
-    op.add_column(
-        "subject",
-        sa.Column(
-            "age_period",
-            postgresql.ENUM("prenatal", "postnatal", name="ageperiod", create_type=False),
-            nullable=True,
-        ),
-    )
-    op.add_column(
-        "subject",
-        sa.Column(
-            "sex", postgresql.ENUM("male", "female", name="sex", create_type=False), nullable=True
-        ),
-    )
-    op.add_column("subject", sa.Column("weight", sa.Float(), nullable=True))
-    op.add_column("subject", sa.Column("description", sa.String(), nullable=False))
-    op.add_column("subject", sa.Column("description_vector", postgresql.TSVECTOR(), nullable=True))
-    op.add_column("subject", sa.Column("species_id", sa.Uuid(), nullable=False))
-    op.add_column("subject", sa.Column("strain_id", sa.Uuid(), nullable=True))
-    op.drop_index("ix_subject_creation_date", table_name="subject")
-    op.drop_index("ix_subject_name", table_name="subject")
-    op.create_index(op.f("ix_subject_name"), "subject", ["name"], unique=False)
-    op.create_index(
-        "ix_subject_description_vector",
-        "subject",
-        ["description_vector"],
-        unique=False,
-        postgresql_using="gin",
-    )
-    op.create_index(op.f("ix_subject_species_id"), "subject", ["species_id"], unique=False)
-    op.create_index(op.f("ix_subject_strain_id"), "subject", ["strain_id"], unique=False)
-    op.create_foreign_key(op.f("fk_subject_id_entity"), "subject", "entity", ["id"], ["id"])
-    op.create_foreign_key(
-        op.f("fk_subject_species_id_species"), "subject", "species", ["species_id"], ["id"]
-    )
-    op.create_foreign_key(
-        "fk_subject_strain_id_species_id",
-        "subject",
-        "strain",
-        ["strain_id", "species_id"],
-        ["id", "species_id"],
-    )
-    op.drop_column("subject", "creation_date")
-    op.drop_column("subject", "update_date")
-    op.sync_enum_values(
-        enum_schema="public",
-        enum_name="entitytype",
-        new_values=[
-            "analysis_software_source_code",
-            "emodel",
-            "experimental_bouton_density",
-            "experimental_neuron_density",
-            "experimental_synapses_per_connection",
-            "memodel",
-            "mesh",
-            "reconstruction_morphology",
-            "single_cell_experimental_trace",
-            "single_neuron_simulation",
-            "single_neuron_synaptome",
-            "single_neuron_synaptome_simulation",
-            "subject",
-        ],
-        affected_columns=[
-            TableReference(table_schema="public", table_name="entity", column_name="type")
-        ],
-        enum_values_to_rename=[],
-    )
-    public_experimental_neuron_density_experimental_neuron_density_description_vector = PGTrigger(
+    public_mesh_mesh_description_vector = PGTrigger(
         schema="public",
-        signature="experimental_neuron_density_description_vector",
-        on_entity="public.experimental_neuron_density",
+        signature="mesh_description_vector",
+        on_entity="public.mesh",
         is_constraint=False,
-        definition="BEFORE INSERT OR UPDATE ON experimental_neuron_density\n            FOR EACH ROW EXECUTE FUNCTION\n                tsvector_update_trigger(description_vector, 'pg_catalog.english', description, name)",
+        definition="BEFORE INSERT OR UPDATE ON mesh\n            FOR EACH ROW EXECUTE FUNCTION\n                tsvector_update_trigger(description_vector, 'pg_catalog.english', description, name)",
     )
-    op.create_entity(
-        public_experimental_neuron_density_experimental_neuron_density_description_vector
-    )
+    op.create_entity(public_mesh_mesh_description_vector)
 
     public_analysis_software_source_code_analysis_software_source_code_description_vector = PGTrigger(
         schema="public",
@@ -204,14 +129,27 @@ def upgrade() -> None:
         public_analysis_software_source_code_analysis_software_source_code_description_vector
     )
 
-    public_mesh_mesh_description_vector = PGTrigger(
+    public_experimental_neuron_density_experimental_neuron_density_description_vector = PGTrigger(
         schema="public",
-        signature="mesh_description_vector",
-        on_entity="public.mesh",
+        signature="experimental_neuron_density_description_vector",
+        on_entity="public.experimental_neuron_density",
         is_constraint=False,
-        definition="BEFORE INSERT OR UPDATE ON mesh\n            FOR EACH ROW EXECUTE FUNCTION\n                tsvector_update_trigger(description_vector, 'pg_catalog.english', description, name)",
+        definition="BEFORE INSERT OR UPDATE ON experimental_neuron_density\n            FOR EACH ROW EXECUTE FUNCTION\n                tsvector_update_trigger(description_vector, 'pg_catalog.english', description, name)",
     )
-    op.create_entity(public_mesh_mesh_description_vector)
+    op.create_entity(
+        public_experimental_neuron_density_experimental_neuron_density_description_vector
+    )
+
+    public_single_cell_experimental_trace_single_cell_experimental_trace_description_vector = PGTrigger(
+        schema="public",
+        signature="single_cell_experimental_trace_description_vector",
+        on_entity="public.single_cell_experimental_trace",
+        is_constraint=False,
+        definition="BEFORE INSERT OR UPDATE ON single_cell_experimental_trace\n            FOR EACH ROW EXECUTE FUNCTION\n                tsvector_update_trigger(description_vector, 'pg_catalog.english', description, name)",
+    )
+    op.create_entity(
+        public_single_cell_experimental_trace_single_cell_experimental_trace_description_vector
+    )
 
     public_experimental_bouton_density_experimental_bouton_density_description_vector = PGTrigger(
         schema="public",
@@ -244,26 +182,6 @@ def upgrade() -> None:
     )
     op.create_entity(public_reconstruction_morphology_reconstruction_morphology_description_vector)
 
-    public_subject_subject_description_vector = PGTrigger(
-        schema="public",
-        signature="subject_description_vector",
-        on_entity="public.subject",
-        is_constraint=False,
-        definition="BEFORE INSERT OR UPDATE ON subject\n            FOR EACH ROW EXECUTE FUNCTION\n                tsvector_update_trigger(description_vector, 'pg_catalog.english', description, name)",
-    )
-    op.create_entity(public_subject_subject_description_vector)
-
-    public_single_cell_experimental_trace_single_cell_experimental_trace_description_vector = PGTrigger(
-        schema="public",
-        signature="single_cell_experimental_trace_description_vector",
-        on_entity="public.single_cell_experimental_trace",
-        is_constraint=False,
-        definition="BEFORE INSERT OR UPDATE ON single_cell_experimental_trace\n            FOR EACH ROW EXECUTE FUNCTION\n                tsvector_update_trigger(description_vector, 'pg_catalog.english', description, name)",
-    )
-    op.create_entity(
-        public_single_cell_experimental_trace_single_cell_experimental_trace_description_vector
-    )
-
     public_reconstruction_morphology_morphology_description_vector = PGTrigger(
         schema="public",
         signature="morphology_description_vector",
@@ -286,26 +204,6 @@ def downgrade() -> None:
         definition="BEFORE INSERT OR UPDATE ON public.reconstruction_morphology FOR EACH ROW EXECUTE FUNCTION tsvector_update_trigger('description_vector', 'pg_catalog.english', 'description', 'name')",
     )
     op.create_entity(public_reconstruction_morphology_morphology_description_vector)
-
-    public_single_cell_experimental_trace_single_cell_experimental_trace_description_vector = PGTrigger(
-        schema="public",
-        signature="single_cell_experimental_trace_description_vector",
-        on_entity="public.single_cell_experimental_trace",
-        is_constraint=False,
-        definition="BEFORE INSERT OR UPDATE ON single_cell_experimental_trace\n            FOR EACH ROW EXECUTE FUNCTION\n                tsvector_update_trigger(description_vector, 'pg_catalog.english', description, name)",
-    )
-    op.drop_entity(
-        public_single_cell_experimental_trace_single_cell_experimental_trace_description_vector
-    )
-
-    public_subject_subject_description_vector = PGTrigger(
-        schema="public",
-        signature="subject_description_vector",
-        on_entity="public.subject",
-        is_constraint=False,
-        definition="BEFORE INSERT OR UPDATE ON subject\n            FOR EACH ROW EXECUTE FUNCTION\n                tsvector_update_trigger(description_vector, 'pg_catalog.english', description, name)",
-    )
-    op.drop_entity(public_subject_subject_description_vector)
 
     public_reconstruction_morphology_reconstruction_morphology_description_vector = PGTrigger(
         schema="public",
@@ -338,24 +236,15 @@ def downgrade() -> None:
         public_experimental_bouton_density_experimental_bouton_density_description_vector
     )
 
-    public_mesh_mesh_description_vector = PGTrigger(
+    public_single_cell_experimental_trace_single_cell_experimental_trace_description_vector = PGTrigger(
         schema="public",
-        signature="mesh_description_vector",
-        on_entity="public.mesh",
+        signature="single_cell_experimental_trace_description_vector",
+        on_entity="public.single_cell_experimental_trace",
         is_constraint=False,
-        definition="BEFORE INSERT OR UPDATE ON mesh\n            FOR EACH ROW EXECUTE FUNCTION\n                tsvector_update_trigger(description_vector, 'pg_catalog.english', description, name)",
-    )
-    op.drop_entity(public_mesh_mesh_description_vector)
-
-    public_analysis_software_source_code_analysis_software_source_code_description_vector = PGTrigger(
-        schema="public",
-        signature="analysis_software_source_code_description_vector",
-        on_entity="public.analysis_software_source_code",
-        is_constraint=False,
-        definition="BEFORE INSERT OR UPDATE ON analysis_software_source_code\n            FOR EACH ROW EXECUTE FUNCTION\n                tsvector_update_trigger(description_vector, 'pg_catalog.english', description, name)",
+        definition="BEFORE INSERT OR UPDATE ON single_cell_experimental_trace\n            FOR EACH ROW EXECUTE FUNCTION\n                tsvector_update_trigger(description_vector, 'pg_catalog.english', description, name)",
     )
     op.drop_entity(
-        public_analysis_software_source_code_analysis_software_source_code_description_vector
+        public_single_cell_experimental_trace_single_cell_experimental_trace_description_vector
     )
 
     public_experimental_neuron_density_experimental_neuron_density_description_vector = PGTrigger(
@@ -369,67 +258,26 @@ def downgrade() -> None:
         public_experimental_neuron_density_experimental_neuron_density_description_vector
     )
 
-    op.sync_enum_values(
-        enum_schema="public",
-        enum_name="entitytype",
-        new_values=[
-            "analysis_software_source_code",
-            "emodel",
-            "experimental_bouton_density",
-            "experimental_neuron_density",
-            "experimental_synapses_per_connection",
-            "memodel",
-            "mesh",
-            "reconstruction_morphology",
-            "single_cell_experimental_trace",
-            "single_neuron_simulation",
-            "single_neuron_synaptome",
-            "single_neuron_synaptome_simulation",
-        ],
-        affected_columns=[
-            TableReference(table_schema="public", table_name="entity", column_name="type")
-        ],
-        enum_values_to_rename=[],
+    public_analysis_software_source_code_analysis_software_source_code_description_vector = PGTrigger(
+        schema="public",
+        signature="analysis_software_source_code_description_vector",
+        on_entity="public.analysis_software_source_code",
+        is_constraint=False,
+        definition="BEFORE INSERT OR UPDATE ON analysis_software_source_code\n            FOR EACH ROW EXECUTE FUNCTION\n                tsvector_update_trigger(description_vector, 'pg_catalog.english', description, name)",
     )
-    op.add_column(
-        "subject",
-        sa.Column(
-            "update_date",
-            postgresql.TIMESTAMP(timezone=True),
-            server_default=sa.text("now()"),
-            autoincrement=False,
-            nullable=False,
-        ),
+    op.drop_entity(
+        public_analysis_software_source_code_analysis_software_source_code_description_vector
     )
-    op.add_column(
-        "subject",
-        sa.Column(
-            "creation_date",
-            postgresql.TIMESTAMP(timezone=True),
-            server_default=sa.text("now()"),
-            autoincrement=False,
-            nullable=False,
-        ),
+
+    public_mesh_mesh_description_vector = PGTrigger(
+        schema="public",
+        signature="mesh_description_vector",
+        on_entity="public.mesh",
+        is_constraint=False,
+        definition="BEFORE INSERT OR UPDATE ON mesh\n            FOR EACH ROW EXECUTE FUNCTION\n                tsvector_update_trigger(description_vector, 'pg_catalog.english', description, name)",
     )
-    op.drop_constraint("fk_subject_strain_id_species_id", "subject", type_="foreignkey")
-    op.drop_constraint(op.f("fk_subject_species_id_species"), "subject", type_="foreignkey")
-    op.drop_constraint(op.f("fk_subject_id_entity"), "subject", type_="foreignkey")
-    op.drop_index(op.f("ix_subject_strain_id"), table_name="subject")
-    op.drop_index(op.f("ix_subject_species_id"), table_name="subject")
-    op.drop_index("ix_subject_description_vector", table_name="subject", postgresql_using="gin")
-    op.drop_index(op.f("ix_subject_name"), table_name="subject")
-    op.create_index("ix_subject_name", "subject", ["name"], unique=True)
-    op.create_index("ix_subject_creation_date", "subject", ["creation_date"], unique=False)
-    op.drop_column("subject", "strain_id")
-    op.drop_column("subject", "species_id")
-    op.drop_column("subject", "description_vector")
-    op.drop_column("subject", "description")
-    op.drop_column("subject", "weight")
-    op.drop_column("subject", "sex")
-    op.drop_column("subject", "age_period")
-    op.drop_column("subject", "age_max")
-    op.drop_column("subject", "age_min")
-    op.drop_column("subject", "age_value")
+    op.drop_entity(public_mesh_mesh_description_vector)
+
     op.drop_index(
         op.f("ix_single_neuron_synaptome_simulation_name"),
         table_name="single_neuron_synaptome_simulation",
@@ -476,6 +324,4 @@ def downgrade() -> None:
         postgresql_using="gin",
     )
     op.drop_column("analysis_software_source_code", "description_vector")
-    sa.Enum("prenatal", "postnatal", name="ageperiod").drop(op.get_bind())
-    sa.Enum("male", "female", name="sex").drop(op.get_bind())
     # ### end Alembic commands ###
