--- alembic/versions/20250415_112543_997bed1c5636_default_migration_message.py
@@ -1,47 +1,0 @@
-"""Default migration message
-
-Revision ID: 997bed1c5636
-Revises: d56333140172
-Create Date: 2025-04-15 11:25:43.222836
-
-"""
-
-from collections.abc import Sequence
-
-from alembic_postgresql_enum import TableReference
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "997bed1c5636"
-down_revision: str | None = "d56333140172"
-branch_labels: str | Sequence[str] | None = None
-depends_on: str | Sequence[str] | None = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.sync_enum_values(
-        enum_schema="public",
-        enum_name="sex",
-        new_values=["male", "female", "unknown"],
-        affected_columns=[
-            TableReference(table_schema="public", table_name="subject", column_name="sex")
-        ],
-        enum_values_to_rename=[],
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.sync_enum_values(
-        enum_schema="public",
-        enum_name="sex",
-        new_values=["male", "female"],
-        affected_columns=[
-            TableReference(table_schema="public", table_name="subject", column_name="sex")
-        ],
-        enum_values_to_rename=[],
-    )
-    # ### end Alembic commands ###
