--- alembic/versions/20250415_114823_7d430f173bf6_default_migration_message.py
@@ -1,47 +1,0 @@
-"""Default migration message
-
-Revision ID: 7d430f173bf6
-Revises: 997bed1c5636
-Create Date: 2025-04-15 11:48:23.958520
-
-"""
-
-from collections.abc import Sequence
-
-from alembic_postgresql_enum import TableReference
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "7d430f173bf6"
-down_revision: str | None = "997bed1c5636"
-branch_labels: str | Sequence[str] | None = None
-depends_on: str | Sequence[str] | None = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.sync_enum_values(
-        enum_schema="public",
-        enum_name="ageperiod",
-        new_values=["prenatal", "postnatal", "unknown"],
-        affected_columns=[
-            TableReference(table_schema="public", table_name="subject", column_name="age_period")
-        ],
-        enum_values_to_rename=[],
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.sync_enum_values(
-        enum_schema="public",
-        enum_name="ageperiod",
-        new_values=["prenatal", "postnatal"],
-        affected_columns=[
-            TableReference(table_schema="public", table_name="subject", column_name="age_period")
-        ],
-        enum_values_to_rename=[],
-    )
-    # ### end Alembic commands ###
