+++ alembic/versions/20250415_160109_8141d8f86f9d_default_migration_message.py
@@ -1,0 +1,47 @@
+"""Default migration message
+
+Revision ID: 8141d8f86f9d
+Revises: d1f17fcb1667
+Create Date: 2025-04-15 16:01:09.849959
+
+"""
+
+from collections.abc import Sequence
+
+from alembic_utils.pg_trigger import PGTrigger
+
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision: str = "8141d8f86f9d"
+down_revision: str | None = "d1f17fcb1667"
+branch_labels: str | Sequence[str] | None = None
+depends_on: str | Sequence[str] | None = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    public_subject_subject_description_vector = PGTrigger(
+        schema="public",
+        signature="subject_description_vector",
+        on_entity="public.subject",
+        is_constraint=False,
+        definition="BEFORE INSERT OR UPDATE ON subject\n            FOR EACH ROW EXECUTE FUNCTION\n                tsvector_update_trigger(description_vector, 'pg_catalog.english', description, name)",
+    )
+    op.create_entity(public_subject_subject_description_vector)
+
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    public_subject_subject_description_vector = PGTrigger(
+        schema="public",
+        signature="subject_description_vector",
+        on_entity="public.subject",
+        is_constraint=False,
+        definition="BEFORE INSERT OR UPDATE ON subject\n            FOR EACH ROW EXECUTE FUNCTION\n                tsvector_update_trigger(description_vector, 'pg_catalog.english', description, name)",
+    )
+    op.drop_entity(public_subject_subject_description_vector)
+
+    # ### end Alembic commands ###
